… CHANGELOG.md
… pygfunction/examples/uniform_temperature.py
… pygfunction/gfunction.py