Closes #171.

Profiling results will be shown below to confirm there was no negative impact on calculation time.