@MassimoCimmino 2021-12-12
## Timing results for `'equivalent'` solver, using `nSegments=8` and default unequal discretization

```
>>> import pygfunction as gt
>>> alpha = 1.0e-6
>>> time = gt.utilities.time_geometric(100*3600., 3000*8760*3600., 25)
>>> method = 'equivalent'
>>> for (N_1, N_2) in [(1, 1), (5, 2), (5, 5), (10, 10), (20, 20)]:
>>>     boreholes = gt.boreholes.rectangle_field(N_1, N_2, 7.5, 7.5, 150., 4., 0.075)
>>>     %timeit -n10 gt.gfunction.gFunction(boreholes, alpha, time=time, method=method)
```

### Before changes

```
499 ms ± 691 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
753 ms ± 17.1 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
753 ms ± 861 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
816 ms ± 975 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.14 s ± 1.83 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### After changes

```
493 ms ± 550 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
736 ms ± 755 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
745 ms ± 774 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
807 ms ± 1.58 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.12 s ± 3.78 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```