@MassimoCimmino 2021-12-12
## Timing results for `'equivalent'` solver, using `nSegments=12` and uniform discretization

```
>>> import pygfunction as gt
>>> alpha = 1.0e-6
>>> time = gt.utilities.time_geometric(100*3600., 3000*8760*3600., 25)
>>> method = 'equivalent'
>>> options = {'nSegments': 12, 'segment_ratios': None}
>>> for (N_1, N_2) in [(1, 1), (5, 2), (5, 5), (10, 10), (20, 20)]:
>>>     boreholes = gt.boreholes.rectangle_field(N_1, N_2, 7.5, 7.5, 150., 4., 0.075)
>>>     %timeit -n10 gt.gfunction.gFunction(boreholes, alpha, time=time, method=method, options=options)
```

### Before changes

```
591 ms ± 578 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
863 ms ± 1.74 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
871 ms ± 742 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
928 ms ± 1.1 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.21 s ± 523 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### After changes

```
590 ms ± 612 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
863 ms ± 1.83 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
871 ms ± 4.41 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
933 ms ± 10.3 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.21 s ± 1.33 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```