@MassimoCimmino 2021-12-12
## Timing results for `'similarities'` solver, using `nSegments=8` and default unequal discretization

```
>>> import pygfunction as gt
>>> alpha = 1.0e-6
>>> time = gt.utilities.time_geometric(100*3600., 3000*8760*3600., 25)
>>> method = 'similarities'
>>> for (N_1, N_2) in [(1, 1), (5, 2), (5, 5), (10, 10)]:
>>>     boreholes = gt.boreholes.rectangle_field(N_1, N_2, 7.5, 7.5, 150., 4., 0.075)
>>>     %timeit -n10 gt.gfunction.gFunction(boreholes, alpha, time=time, method=method)
```

### Before changes

```
421 ms ± 809 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
729 ms ± 1.03 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
804 ms ± 3.95 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.95 s ± 14.7 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### After changes

```
428 ms ± 9.75 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
726 ms ± 1.22 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
803 ms ± 3.64 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.96 s ± 12.7 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```