@MassimoCimmino 2021-12-12
## Timing results for `'similarities'` solver, using `nSegments=12` and uniform discretization

```
>>> import pygfunction as gt
>>> alpha = 1.0e-6
>>> time = gt.utilities.time_geometric(100*3600., 3000*8760*3600., 25)
>>> method = 'similarities'
>>> options = {'nSegments': 12, 'segment_ratios': None}
>>> for (N_1, N_2) in [(1, 1), (5, 2), (5, 5), (10, 10)]:
>>>     boreholes = gt.boreholes.rectangle_field(N_1, N_2, 7.5, 7.5, 150., 4., 0.075)
>>>     %timeit -n10 gt.gfunction.gFunction(boreholes, alpha, time=time, method=method, options=options)
```

### Before changes

```
484 ms ± 1.89 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
855 ms ± 906 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
989 ms ± 5.34 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
3.37 s ± 12 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### After changes

```
485 ms ± 580 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
860 ms ± 975 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
995 ms ± 4.69 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
3.37 s ± 8.38 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```